fix: clippy lints across the codebase
- Remove unnecessary cast in `fec.rs`
- Use `is_multiple_of` for parity checks in `resample.rs`
- Replace needless range loop with iterator in `zadoff_chu.rs`
- Replace useless `vec!` with slice in `equalization.rs` tests
- Refactor `loop` with `if let` to `while let` in `generate-waveform.rs`

+++ dsp/src/coding/fec.rs
@@ -284,7 +284,7 @@ pub fn decode_soft_list(llrs: &[f32], list_size: usize) -> Vec<Vec<u8>> {
                 decoded[t] = sv.bit;
             }
             state = sv.prev_state as usize;
-            rank = sv.prev_rank as usize;
+            rank = sv.prev_rank;
         }
 
         if valid_path && !outputs.iter().any(|v| v == &decoded) {
